requirements.txt
pymongo
xdelta3
boto3
flask
psycopg2-binary
python-dotenv